updated the test case class and added the logic to compare the temp values and created the Readme file

+++ src/main/java/com/weatherInfo/pages/WeatherAPIResponse.java
@@ -22,7 +22,7 @@ public class WeatherAPIResponse {
 	public BasePage basepage;
 	
 	
-	public int getWeatherAPI() {
+	public double getWeatherAPI() {
 		
 		
 		/*
@@ -45,7 +45,7 @@ public class WeatherAPIResponse {
 		String temp = json.getString(".main.temp");
 		System.out.println(temp);
 		
-		int tempurature = Integer.parseInt(temp);
+		double tempurature = Double.parseDouble(temp);
 		
 		return tempurature;
